--Major Changes--
Changed to Lazy.nvim for plugin manager which is more performant than
Packer and also separated plugin's config and build related
configuration for better modularity in the future and to be able to
clearly distinguish where the plugins are. Also in the init.lua file
I've added the Lazy config with the loading of the plugins lazily to
load them just when necessary, removed old packer files and overall
better plugin management. :)